fix(rktio): module dependencies btw generated and manual bindings
